still reachable leaks solved in heredocs

+++ src/free_values.c
@@ -6,7 +6,7 @@
 /*   By: vnafissi <vnafissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/11 19:23:39 by vladimir          #+#    #+#             */
-/*   Updated: 2022/07/23 19:13:35 by vnafissi         ###   ########.fr       */
+/*   Updated: 2022/07/23 19:53:30 by vnafissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,14 +25,17 @@ void	free_token_lst(t_token *token)
 	}
 }
 
-void	free_delete_heredoc_file(t_cmd_line *cmd_line)
+void	free_delete_heredoc_file(t_cmd_line *cmd_line, int del_hd_file)
 {
-	if (cmd_line->heredoc_mode == 1)
-		unlink(cmd_line->heredoc_name);
+	if (del_hd_file)
+	{
+		if (cmd_line->heredoc_mode == 1)
+			unlink(cmd_line->heredoc_name);
+	}
 	ft_free_null_str(&cmd_line->heredoc_name);
 }
 
-void	free_cmd_line_lst(t_cmd_line *cmd_line)
+void	free_cmd_line_lst(t_cmd_line *cmd_line, int del_hd_file)
 {
 	t_cmd_line	*next;
 
@@ -41,7 +44,7 @@ void	free_cmd_line_lst(t_cmd_line *cmd_line)
 		next = cmd_line->next;
 		free_token_lst(cmd_line->token_lst);
 		ft_free_null_str(&cmd_line->cmd);
-		free_delete_heredoc_file(cmd_line);
+		free_delete_heredoc_file(cmd_line, del_hd_file);
 		ft_freetbl(cmd_line->args, -1);
 		if (cmd_line->fdin != 0 && cmd_line->fdin != -1)
 			close(cmd_line->fdin);
@@ -53,18 +56,18 @@ void	free_cmd_line_lst(t_cmd_line *cmd_line)
 	}
 }
 
-void	free_values(t_sh *sh, int free_env)
+void	free_values(t_sh *sh, int free_env, int del_hd_file)
 {
 	free_token_lst(sh->token_lst);
 	ft_free_null_str(&sh->prompt);
-	free_cmd_line_lst(sh->cmd_line_lst);
+	free_cmd_line_lst(sh->cmd_line_lst, del_hd_file);
 	if (free_env)
 		ft_freetbl(sh->env, -1);
 }
 
-void	ft_free_values_exit(t_sh *sh, int exit_status)
+void	ft_free_values_exit(t_sh *sh, int exit_status, int del_hd_file)
 {
-	free_values(sh, 1);
+	free_values(sh, 1, del_hd_file);
 	rl_clear_history();
 	exit(exit_status);
 }
